+++ src/QueryCat.Backend/Commands/Insert/InsertCommand.cs
@@ -20,7 +20,7 @@ internal class InsertCommand : ICommand
         var rowsOutput = new CreateDelegateVisitor(executionThread)
             .RunAndReturn(insertNode.InsertTargetNode)
             .Invoke()
-            .GetAsObject<IRowsOutput>();
+            .As<IRowsOutput>();
 
         // Evaluate iterator for FROM block and get input source.
         var outputDefinedColumns = insertNode.HasDefinedColumns();